+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - MIG-1039 - [BREAKING] Added new method `getCountryStateUuid` to `SwagMigrationAssistant\Migration\Mapping\MappingServiceInterface`
 - MIG-1039 - [BREAKING] Added new constructor parameter `EntityRepository $countryStateRepo` to `SwagMigrationAssistant\Migration\Mapping\MappingService`
 - MIG-1039 - [BREAKING] Removed method `getCountryStateUuid` from `SwagMigrationAssistant\Profile\Shopware6\Mapping\Shopware6MappingServiceInterface` and its default implementation `SwagMigrationAssistant\Profile\Shopware6\Mapping\Shopware6MappingService`
+- MIG-1050 - Improves the migration of SEO URLs by using the `SeoUrlPersister::updateSeoUrls` method
 
 # 13.2.0
 - MIG-1035 - Changed the premapping assignment in the administration, which should fix issues with assignment on later pagination pages and changes getting lost with newly selected data